ref: Some refactorings
