+++ packages/browser/src/lib/frontend.ts
@@ -1,15 +1,6 @@
-import { createAndBind, FrontendBase, SdkInfo } from '@sentry/core';
+import { FrontendBase, SdkInfo } from '@sentry/core';
 import { BrowserBackend, BrowserOptions } from './backend';
 import { Raven } from './raven';
-export { addBreadcrumb, captureEvent, setUserContext } from '@sentry/core';
-export {
-  captureException,
-  captureMessage,
-  popScope,
-  pushScope,
-  setExtraContext,
-  setTagsContext,
-} from '@sentry/shim';
 
 /**
  * The Sentry Browser SDK Frontend.
@@ -50,48 +41,4 @@ export class BrowserFrontend extends FrontendBase<
   public wrap(fn: Function, options: object): Function {
     return Raven.wrap(options, fn);
   }
-}
-
-/**
- * The Sentry Browser SDK Client.
- *
- * To use this SDK, call the {@link Sdk.create} function as early as possible
- * when loading the web page. To set context information or send manual events,
- * use the provided methods.
- *
- * @example
- * const { SentryClient } = require('@sentry/browser');
- *
- * SentryClient.create({
- *   dsn: '__DSN__',
- *   // ...
- * });
- *
- * @example
- * SentryClient.setContext({
- *   extra: { battery: 0.7 },
- *   tags: { user_mode: 'admin' },
- *   user: { id: '4711' },
- * });
- *
- * @example
- * SentryClient.addBreadcrumb({
- *   message: 'My Breadcrumb',
- *   // ...
- * });
- *
- * @example
- * SentryClient.captureMessage('Hello, world!');
- * SentryClient.captureException(new Error('Good bye'));
- * SentryClient.captureEvent({
- *   message: 'Manual',
- *   stacktrace: [
- *     // ...
- *   ],
- * });
- * TODO
- * @see BrowserOptions for documentation on configuration options.
- */
-export function create(options: BrowserOptions): void {
-  createAndBind(BrowserFrontend, options);
 }
