+++ packages/core/src/lib/domain.ts
@@ -117,9 +117,3 @@ export interface SentryEvent extends Context {
   stacktrace?: Stacktrace;
   breadcrumbs?: Breadcrumb[];
 }
-
-/** TODO */
-export interface Scope {
-  breadcrumbs: Breadcrumb[];
-  context: Context;
-}
